Consolidate wine code

+++ platform/windows/export/export_plugin.cpp
@@ -424,20 +424,6 @@ Error EditorExportPlatformWindows::_rcedit_add_data(const Ref<EditorExportPreset
 		rcedit_path = "rcedit"; // try to run rcedit from PATH
 	}
 
-#ifndef WINDOWS_ENABLED
-	// On non-Windows we need WINE to run rcedit
-	String wine_path = EDITOR_GET("export/windows/wine");
-
-	if (!wine_path.is_empty() && !FileAccess::exists(wine_path)) {
-		add_message(EXPORT_MESSAGE_WARNING, TTR("Resources Modification"), vformat(TTR("Could not find wine executable at \"%s\"."), wine_path));
-		return ERR_FILE_NOT_FOUND;
-	}
-
-	if (wine_path.is_empty()) {
-		wine_path = "wine"; // try to run wine from PATH
-	}
-#endif
-
 	String icon_path;
 	if (p_preset->get("application/icon") != "") {
 		icon_path = p_preset->get("application/icon");
@@ -513,11 +499,24 @@ Error EditorExportPlatformWindows::_rcedit_add_data(const Ref<EditorExportPreset
 	}
 
 #ifndef WINDOWS_ENABLED
-	// On non-Windows we need WINE to run rcedit
-	args.push_front(rcedit_path);
-	rcedit_path = wine_path;
-	OS::get_singleton()->set_environment("WINEDEBUG", "-all"); // Disable wine debug messages
-	OS::get_singleton()->set_environment("WINEDLLOVERRIDES", "mscoree="); // Prevent wine-mono install prompt on first run
+	// On non-Windows we need WINE to run rcedit unless its being run via PATH
+	if (rcedit_path != "rcedit") {
+		String wine_path = EDITOR_GET("export/windows/wine");
+
+		if (!wine_path.is_empty() && !FileAccess::exists(wine_path)) {
+			add_message(EXPORT_MESSAGE_WARNING, TTR("Resources Modification"), vformat(TTR("Could not find wine executable at \"%s\"."), wine_path));
+			return ERR_FILE_NOT_FOUND;
+		}
+
+		if (wine_path.is_empty()) {
+			wine_path = "wine"; // try to run wine from PATH
+		}
+
+		args.push_front(rcedit_path);
+		rcedit_path = wine_path;
+		OS::get_singleton()->set_environment("WINEDEBUG", "-all"); // Disable wine debug messages
+		OS::get_singleton()->set_environment("WINEDLLOVERRIDES", "mscoree="); // Prevent wine-mono install prompt on first run
+	}
 #endif
 
 	String str;
